+++ test/src/Game/Werewolf/Test/Command.hs
@@ -105,6 +105,7 @@ allCommandTests =
     , testProperty "quit command clears protect when caller is defender"        prop_quitCommandClearsProtectWhenCallerIsDefender
     , testProperty "quit command clears player's devour vote"                   prop_quitCommandClearsPlayersDevourVote
     , testProperty "quit command clears player's lynch vote"                    prop_quitCommandClearsPlayersLynchVote
+    , testProperty "quit command clears role model when caller is wild-child"   prop_quitCommandClearsRoleModelWhenCallerIsWildChild
 
     , testProperty "see command errors when game is over"           prop_seeCommandErrorsWhenGameIsOver
     , testProperty "see command errors when caller does not exist"  prop_seeCommandErrorsWhenCallerDoesNotExist
@@ -584,6 +585,13 @@ prop_quitCommandClearsPlayersLynchVote (GameWithLynchVotes game) =
         let command = quitCommand (caller ^. name)
 
         isNothing $ run_ (apply command) game ^. votes . at (caller ^. name)
+
+prop_quitCommandClearsRoleModelWhenCallerIsWildChild :: GameWithRoleModel -> Bool
+prop_quitCommandClearsRoleModelWhenCallerIsWildChild (GameWithRoleModel game) = do
+    let wildChild   = findByRole_ wildChildRole (game ^. players)
+    let command     = quitCommand (wildChild ^. name)
+
+    isNothing $ run_ (apply command) game ^. roleModel
 
 prop_seeCommandErrorsWhenGameIsOver :: GameAtGameOver -> Property
 prop_seeCommandErrorsWhenGameIsOver (GameAtGameOver game) =